Refactored dataset loading and evaluation into a separate module, updated mlora_train.py
